+++ InventoryTools/Logic/Columns/QuantityColumn.cs
@@ -117,7 +117,7 @@ namespace InventoryTools.Logic.Columns
             return base.Draw(configuration, columnConfiguration, searchResult, rowIndex, columnIndex);
         }
 
-        public override string Name { get; set; } = "Total Quantity Available";
+        public override string Name { get; set; } = "Quantity/Total Quantity Available";
         public override string RenderName => "Quantity";
 
         public override float Width { get; set; } = 70.0f;